разделение кода на модули, описание каждого модуля
